README.md
I am following the manual found http://anavi.technology/files/anavi-infrared-phat.pdf
It has 2 glaring errors on pages 5 & 6. If you examine the install commands they are incomplete. There is obviously more to install on both pages, the cut and paste action used in the creation of the pdf file cut off some of the text, because there were too many characters to fit in one line of the pdf. This why there is an slide bar under the line of text.
All though it doesn’t seem to effect the operation of the lirc software, my OCD forces me ask if anyone has the full “uncut” version of those two installs